backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.14.4",
    "fastapi[standard]>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.3.0",
    "langgraph>=1.2.0",
    "openai>=2.36.0",
    "psycopg[binary]>=3.3.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.47.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.1.0",
    "langchain-tavily>=0.2.18",
]

[dependency-groups]
dev = [
    "pyright>=1.1.409",
    "pytest>=9.0.3",
    "ruff>=0.15.13",
]
